requirements.txt
aiosqlite==0.20.0
playwright==1.42.0
groq==0.4.2
python-dotenv==1.0.1
